Refactor: use one input channel for subwindow
Originally, subwindows are connected by doubly-linked channels. Each
subwindow has input from left and right, and output to left and right.
This is not necessary, as the subwindows may have different connection
diagrams in handshake join and broadcast join.

Therefore, we can use a single input channel for each subwindow,
which contains all the input from the outside world.