Get next tuple in chronological order in broadcast join master
Originally, the master thread decides how to get the next tuple from the 2 streams interleavingly. If the arriving rate is different between R and S, the channel order will depart from the global timestamp order.

Now, the tuple reader is a reader that can handle chronologically correct reading from 1/2 streams as long as the stream tuples has incremental timestamps and reasonable available-to-read behavior