Support multiple workers using watchdog for input
The watchdog (check close thread) periodically reset a flag, while it is set every time a new tuple is read.

If the watchdog sees a unsetted flag, it assumes the input channel is empty for a long time and no tuples will come. Then the process routine can be terminated.