Use pending tuples list to avoid channel full with no input control.
Originally the tuple reader has to limit the reading rate to avoid
making the channels full, which could probably cause a deadlock
situation. This is not a good solution, because the input streams should
be able to read as fast as possible.

The solution is to use a pending list of tuples. Every time a tuple is to
be sent to the channel, it is added to the pending list. And the main loop
will check if the pending list is empty and send the tuples stored in the
list when the underlying output channels are not quite full.

This requires a change to read tuples from the input channel: from reading
one tuple blockingly to check whether the channel is empty or not.