Fix: Add forwarded restriction to avoid duplicate results
Add validate_join.sh script to check for unmatched results in the 2 join
diagrams, especially for duplicates and missing ones in handshake join.

In the pending list implementation, the "r" tuple is forwarded logically
to the right pending list but not physically deleted from the index_r.
So an "s" tuple could join with 2 identical "r" tuples in the current
window and right neighbor window. To avoid this, we need to add a restriction
to the join condition to check if the "r" tuple is already in the pending
list, which is logically deleted.