Fix: Force the producer to push tuples to the windows to avoid deadlocks
In random streams, the duplicated key tuples will not be inserted into the
window, which may cause the end windows to have a relatively new tuple
at the top of the queue. But the producer still waits for the end window
to pop the tuple to push the new tuples inside, which causes a deadlock.

Solution: Force the producer to push tuples to the windows after failing
too many times.