Use 2 producer routines to push in handshake
This commit finishes the sliding window source control for the handshake joiner.
The sliding window source control is a mechanism to control the tuple
insertion and deletion in the 2 end windows. The producer routines
push tuples when it is much newer than the oldest tuple in the end window.
The end window pops the expired tuples which will not match with the incoming
tuples in the opposite stream. This mechanism is deadlock-prune and an extra
window threshold (PUSH_TUPLE_TOLERANCE) is necessary for the producer to
push a little bit more tuples given the timestamp condition.

The main bug of the deadlock is duplicate tuple insertion. When the end
window wants to pop its oldest tuple, this tuple may be far newer than the
previous popped tuple (oldest_ts).