Add AlexMap with false negative find bug
Alex Map is a learned index that supports fast range search and insert/erase. But in the unit test, alex map seldomly fails to find an inserted key which causes deadlock in handshake join.

But given the config.yaml, the current configuration will not cause main program to crash when using alex map as index. However, this is a potential danger if using more tuples and larger window size in the benchmark test.