Fix: Move shuffle logic from load to construction
Previously, the shuffle happens during the load_sosd() function. If 2 sosd streams are created from the main.cpp, then they always have exactly the same tuple order, which is unexpected.

This commit move the shuffle logic to the sosd stream constructor with an input seed. If seed is not given, we randomly create a new seed and then shuffle